flkjdslkfjl no recuerdo los cambios

+++ api/controllers/things/destroy-one-thing.js
@@ -16,11 +16,22 @@ module.exports = {
 
 
     exits: {
-
+        forbidden: {
+            description: 'El usuario esta hacendo este request no tiene permisos para borrar esto',
+            responseType: 'forbidden'
+        }
     },
 
 
     fn: async function(inputs, exits) {
+
+        var thing = await Thing.findOne({
+            id: inputs.id
+        });
+
+        if (thing.owner !== this.req.me.id) {
+            throw 'forbidden';
+        }
 
         await Thing.destroy({ id: inputs.id });
